Catch exception when sorting keys for presentation
When printing keys (with debug turned on), we try to sort the keys
to make it easier to read.  But keys may not all be strings.  So,
try to sort, and if that fails, abandon the sort.

(Better strategy, for the future: separate the keys into different
lists by type, then sort those lists independently, then concatenate
them all together.)